Move navbar-header component to top-navbar dir
