+++ app/javascript/packs/component-definitions-common.js
@@ -24,7 +24,7 @@ import ImportDatastoreViaGit from '../components/automate-import-export-form/imp
 import MainMenu from '../components/main-menu/main-menu';
 import MiqAboutModal from '../components/miq-about-modal';
 import MiqToolbar from '../components/miq-toolbar';
-import NavbarHeader from '../components/navbar-header/navbar-header';
+import NavbarHeader from '../components/top-navbar/navbar-header';
 import navbar from '../components/top-navbar';
 import OptimizationList from '../optimization/optimization_list';
 import OpsTenantForm from '../components/ops-tenant-form/ops-tenant-form';
